Remove `process.env` refs from the browser artifact
Both the Node.js and the browser artifacts that we generate contain
dynamic references to entries in `process.env`. This does not play
nicely when bundling apps intended for the browser.

As of now, the only function that does this is `getBaseUrl`, which is
a function intended for internal development only. This change makes
sure that this function is defined for the browser in such a way that
does not reference `process.env` anymore and just returns the
Production base URL.